Print the URL and open an iframe if in notebook
