+++ dagshub/data_engine/model/datasource.py
@@ -21,6 +21,7 @@ import dagshub.common.config
 from dagshub.common import rich_console
 from dagshub.common.analytics import send_analytics_event
 from dagshub.common.helpers import prompt_user, http_request, log_message
+from dagshub.common.notebook import open_notebook_iframe
 from dagshub.common.rich_util import get_rich_progress
 from dagshub.common.util import lazy_load, multi_urljoin, to_timestamp, exclude_if_none
 from dagshub.data_engine.client.models import (
@@ -836,9 +837,16 @@ class Datasource:
         Read the function docs for kwarg documentation.
         """
         if visualizer == "dagshub":
-            url = self._generate_visualize_url()
-            webbrowser.open(url)
-            return url
+            link = self._generate_visualize_url()
+            webbrowser.open(link)
+
+            print("The visualization is available at the following link:")
+            print(link)
+
+            open_notebook_iframe(link)
+
+            return link
+
         elif visualizer == "fiftyone":
             return self.all().visualize(**kwargs)
